refactor: use correct feature name for backend classes
